// apps/desktop/src/operations/mutations/view-create.ts
export type ViewCreateMutationInput = {
  type: 'view_create';
  userId: string;
  databaseId: string;
  viewType: 'table' | 'board' | 'calendar';
  name: string;
  groupBy: string | null;
};

export type ViewCreateMutationOutput = {
  id: string;
};

declare module '@/operations/mutations' {
  interface MutationMap {
    view_create: {
      input: ViewCreateMutationInput;
      output: ViewCreateMutationOutput;
    };
  }
}